Preparing for release AKA actually testing stuff * Fixed string/int issue with mission timeframes * Fixed string/int issue with quest cooldowns * Fixed saving missions that referenced missing worlds * Fixed not all missions being loaded into the viewer * Fixed cooldowns/status * Fixed minutes display in Text.timeFromNum * Fixed missing config settings on upgrade * Fixed sounds (as best as possible) on upgrade * Fixed editor error when displaying quests/categories with many elements inside * Fixed "has questbook" check force creating a PlayerManager immediately on join

+++ src/main/java/me/mrCookieSlime/QuestWorld/api/contract/IQuest.java
@@ -15,8 +15,8 @@ public interface IQuest extends IStateful {
 	ICategory getCategory();
 	List<ItemStack> getRewards();
 	IMission getMission(int i);
-	long getRawCooldown();
-	long getCooldown();
+	int getRawCooldown();
+	int getCooldown();
 	int getMoney();
 	int getPartySize();
 	int getXP();
